Rework compiling the gem
Use MiniPortile gem instead of calling configure, make and copying files
to `./lib` manually.

This means `lib/xapian.rb` is always updated when running `rake compile`
so we don't have to manually update. We could remove this file but it is
a good guide to see if the interface has changed upstream.